Renamed things to make them more readable.

+++ src/main/java/frc/robot/Constants.java
@@ -67,5 +67,5 @@ public final class Constants {
     public static final int backRightEncoder = 1;
     public static final int backLeftEncoder = 2;
 
-    public static final double[] turnEncoderOffsets = {256, 346, 171, 95};
+    public static final int[] turnEncoderOffsets = {256, 346, 171, 95};
 }
